WindowPadX:  Adds info on non-primary monitors
AutoHotkey does not have per-monitor DPI awarenes.
For that reason, window placement will not work on monitors that have different scaling than the primary monitor.

(Maybe it would be posible to use two hotkeys to set 'scaling' for a monitor, and thus have the program use that value instead.
Or having it stored in config connected to monitors ID or serialnumber etc.)